@@ -1,49 +1,47 @@
 {
-  "name": "sailing-to-mars-dj-site",
+  "name": "gatsby-ts",
   "version": "1.0.0",
   "private": true,
-  "description": "Sailing To Mars DJ Site",
-  "author": "Spenser Saling",
+  "description": "gatsby-ts",
+  "author": "spensbot",
   "keywords": [
     "gatsby"
   ],
   "scripts": {
-    "dev": "gatsby develop",
+    "develop": "gatsby develop",
     "start": "gatsby develop",
     "build": "gatsby build",
     "serve": "gatsby serve",
-    "clean": "gatsby clean"
+    "clean": "gatsby clean",
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@date-io/dayjs": "^1.3.13",
-    "@fontsource/pacifico": "^4.5.0",
-    "@fontsource/roboto": "^4.5.0",
-    "@fontsource/urbanist": "^4.5.1",
-    "@googlemaps/js-api-loader": "^1.12.2",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/pickers": "^3.3.10",
-    "@types/react": "^17.0.20",
-    "@types/react-beautiful-dnd": "^13.1.2",
-    "@types/styled-components": "^5.1.14",
-    "babel-plugin-styled-components": "^1.13.2",
-    "dayjs": "^1.10.6",
-    "gatsby": "^3.13.1",
-    "gatsby-plugin-image": "^1.13.0",
-    "gatsby-plugin-manifest": "^3.13.0",
-    "gatsby-plugin-portal": "^1.0.7",
-    "gatsby-plugin-prefetch-google-fonts": "^1.4.3",
-    "gatsby-plugin-react-helmet": "^4.13.0",
-    "gatsby-plugin-sharp": "^3.13.0",
-    "gatsby-plugin-sitemap": "^4.9.0",
-    "gatsby-plugin-styled-components": "^4.13.0",
-    "gatsby-plugin-web-font-loader": "^1.0.4",
-    "gatsby-source-filesystem": "^3.13.0",
-    "gatsby-transformer-remark": "^4.10.0",
-    "gatsby-transformer-sharp": "^3.13.0",
-    "react": "^17.0.1",
-    "react-beautiful-dnd": "^13.1.0",
-    "react-dom": "^17.0.1",
-    "react-helmet": "^6.1.0",
-    "styled-components": "^5.3.1"
+    "@mdx-js/react": "^2.1.5",
+    "@mui/material": "^5.10.13",
+    "babel-plugin-styled-components": "^2.0.7",
+    "date-fns": "^2.29.3",
+    "gatsby": "^5.0.0",
+    "gatsby-plugin-google-analytics": "^5.0.0",
+    "gatsby-plugin-image": "^3.0.0",
+    "gatsby-plugin-manifest": "^5.0.0",
+    "gatsby-plugin-mdx": "^5.0.0",
+    "gatsby-plugin-remove-serviceworker": "^1.0.0",
+    "gatsby-plugin-sharp": "^5.0.0",
+    "gatsby-plugin-sitemap": "^6.0.0",
+    "gatsby-plugin-styled-components": "^6.0.0",
+    "gatsby-source-filesystem": "^5.0.0",
+    "gatsby-transformer-sharp": "^5.0.0",
+    "react": "^18.2.0",
+    "react-day-picker": "^8.3.6",
+    "react-dom": "^18.2.0",
+    "react-flatpickr": "^3.10.13",
+    "styled-components": "^5.3.6"
+  },
+  "devDependencies": {
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.8",
+    "@types/react-flatpickr": "^3.8.8",
+    "typescript": "^4.8.4"
   }
 }
