@@ -1,7 +1,6 @@
 import * as React from "react"
-import { Link } from "gatsby"
+import { Link, HeadFC, PageProps } from "gatsby"
 
-// styles
 const pageStyles = {
   color: "#232129",
   padding: "96px",
@@ -24,18 +23,12 @@ const codeStyles = {
   borderRadius: 4,
 }
 
-// markup
-const NotFoundPage = () => {
+const NotFoundPage: React.FC<PageProps> = () => {
   return (
     <main style={pageStyles}>
-      <title>Not found</title>
       <h1 style={headingStyles}>Page not found</h1>
       <p style={paragraphStyles}>
-        Sorry{" "}
-        <span role="img" aria-label="Pensive emoji ">
-          😔
-        </span>{" "}
-        we couldn’t find what you were looking for.
+        Sorry 😔, we couldn’t find what you were looking for.
         <br />
         {process.env.NODE_ENV === "development" ? (
           <>
@@ -52,3 +45,5 @@ const NotFoundPage = () => {
 }
 
 export default NotFoundPage
+
+export const Head: HeadFC = () => <title>Not found</title>
